LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-auto-text-fragment-expected.txt


FAIL Fragment navigation with content-visibility; single text assert_equals: expected "text" but got "top"
FAIL Fragment navigation with content-visibility; range across blocks assert_equals: expected "text2" but got "top"
